Candidate charities to be presented to the user.
Added the logic to pick the candidate charities to be presented
to users. Some charities have good metadata (like a mission statement,
cause, tagline etc.) whereas others are basically a name and
an address.

For that reason, we don't want to randomly pick a bunch of charities
but want to bias towards charities with good metadata. At the same time,
we also don't want to completely exclude charities without great metadata.

So, in this commit, we added the logic to load the candidate charities
that will be biased towards charities with good metadata. There is also
logic to make sure we don't keep calculating the list of candidate charities
on every user request. Currenlty, it is configured to calcualte candidate
charities for about 5% of the requests.
(Note, candidate charities persist across user sessions and so all requests
after candidate charities are reevaluated get the new list).

Also, we currently limit the candidate charities list to 50 which should
be more than enough. Hardly anyone is going to listed through 50 charities.