Data management code.
Committing a bunch of data management code for Alexa charities.
There is a data file that simply holds the list of charities that
accept donations through Alexa. Then there are scripts that search for
these charities on Charity Navigator and build a data file with all the
details for Alexa supported charities.

While searching on Charity Navigator, we create two files. One with high
confidence matches where we found the exact charity. Another with low
confidence matches where we generate five close matches and then requires
a manual review to pick the right one.